* refs/remotes/origin/main: (1809 commits)
  Bump the gradle-deps group across 1 directory with 5 updates
  Also depend on `checker-qual` artifact
  Bump the github-actions-deps group with 5 updates
  Minor refactoring to remove duplicate code
  fixing the broken automode
  resolve reviewer requests
  applied formatting style
  disable automatic formatting of Java code blocks in comments/JavaDoc
  removed formatter version lock and added new keys (via new styleMerge tool)
  added small utility to merge xml formatter style files
  add javadoc
  add javadoc
  reformat fix finalize() deprecation
  fix error in expecting proof script if there is none
  fix compile errors due to merging
  forgot ProofScriptEntry
  fix compile errors
  Removal of the Triple class
  Removal of Quadruple.java
  fix auto merger in github workflow
  ...

# Conflicts:
#	key.core/src/main/java/de/uka/ilkd/key/control/KeYEnvironment.java
#	key.core/src/main/java/de/uka/ilkd/key/macros/scripts/AllCommand.java
#	key.core/src/main/java/de/uka/ilkd/key/macros/scripts/AutoCommand.java
#	key.core/src/main/java/de/uka/ilkd/key/macros/scripts/EngineState.java
#	key.core/src/main/java/de/uka/ilkd/key/macros/scripts/HideCommand.java
#	key.core/src/main/java/de/uka/ilkd/key/macros/scripts/ProofScriptEngine.java
#	key.core/src/main/java/de/uka/ilkd/key/macros/scripts/SaveInstCommand.java
#	key.core/src/main/java/de/uka/ilkd/key/macros/scripts/UnhideCommand.java
#	key.core/src/main/java/de/uka/ilkd/key/macros/scripts/meta/ValueInjector.java
#	key.core/src/main/java/de/uka/ilkd/key/nparser/KeyAst.java
#	key.core/src/main/java/de/uka/ilkd/key/nparser/ParsingFacade.java
#	key.core/src/main/java/de/uka/ilkd/key/proof/init/KeYUserProblemFile.java
#	key.core/src/main/java/de/uka/ilkd/key/proof/io/AbstractProblemLoader.java
#	key.core/src/main/java/de/uka/ilkd/key/speclang/PositionedString.java
#	key.core/src/main/java/de/uka/ilkd/key/speclang/njml/JmlFacade.java
#	key.core/src/test/java/de/uka/ilkd/key/logic/TestLocalSymbols.java
#	key.core/src/test/java/de/uka/ilkd/key/macros/scripts/TestProofScriptCommand.java
#	key.core/src/test/java/de/uka/ilkd/key/proof/proverules/ProveRulesTest.java
#	key.core/src/test/java/de/uka/ilkd/key/proof/runallproofs/ProveTest.java
#	key.core/src/test/java/de/uka/ilkd/key/proof/runallproofs/performance/DataRecordingTestFile.java
#	key.core/src/test/java/de/uka/ilkd/key/proof/runallproofs/proofcollection/TestFile.java
#	key.core/tacletProofs/seqPerm2/Taclet_schiffl_lemma_2.proof
#	key.ui/src/main/java/de/uka/ilkd/key/gui/ProofScriptWorker.java
#	key.ui/src/main/java/de/uka/ilkd/key/gui/WindowUserInterfaceControl.java
#	key.ui/src/main/java/de/uka/ilkd/key/gui/actions/ProofScriptInputAction.java
#	key.ui/src/main/java/de/uka/ilkd/key/ui/ConsoleUserInterfaceControl.java

+++ key.core/src/main/java/de/uka/ilkd/key/macros/scripts/AssertCommand.java
@@ -1,3 +1,6 @@
+/* This file is part of KeY - https://key-project.org
+ * KeY is licensed under the GNU General Public License Version 2
+ * SPDX-License-Identifier: GPL-2.0-only */
 package de.uka.ilkd.key.macros.scripts;
 
 import java.util.Map;
@@ -31,9 +34,9 @@ public class AssertCommand extends AbstractCommand<AssertCommand.Parameters> {
             throw new ScriptException("No parameter specified!");
         }
 
-        if (state.getProof().openEnabledGoals().size() != args.goals.intValue()) {
+        if (state.getProof().openEnabledGoals().size() != args.goals) {
             throw new ScriptException("Assertion failed: number of open goals is "
-                + state.getProof().openGoals().size() + ", but should be " + args.goals.intValue());
+                + state.getProof().openGoals().size() + ", but should be " + args.goals);
         }
     }
