+++ key.ui/src/main/java/de/uka/ilkd/key/ui/Verbosity.java
@@ -1,9 +1,12 @@
+/* This file is part of KeY - https://key-project.org
+ * KeY is licensed under the GNU General Public License Version 2
+ * SPDX-License-Identifier: GPL-2.0-only */
 package de.uka.ilkd.key.ui;
 
 public interface Verbosity {
-    byte SILENT = 0;
-    byte NORMAL = 1;
-    byte INFO = 2;
-    byte DEBUG = 3;
-    byte TRACE = 4;
+    int SILENT = 0;
+    int NORMAL = 1;
+    int INFO = 2;
+    int DEBUG = 3;
+    int TRACE = 4;
 }
